Added real-case scenario added resource server using same auth provier
