+++ README.md
@@ -69,6 +69,29 @@ npm start
 - Note: Angular6 modules are commented out because these don't build on Jenkins as they need npm installed, but they build properly locally
 - Note for Angular version < 4.3.0: You should comment out the HttpClient and HttpClientModule import in app.module and app.service.ts. These version rely on the HttpModule.
 
+## Using the JS-only SPA OAuth Client
+The main purpose of these projects are to analyze how OAuth should be carried out on Javascript-only Single-Page-Applications, using the authorization_code flow with PKCE.
+
+The *SPA/clients-js-only-react* project includes a very simple Spring Boot Application serving a couple of separate Single-Page-Applications developed in React.
+
+It includes two pages:
+  * a 'Step-By-Step' guide, where we analyze explicitly each step that we need to carry out to obtain an access token and request a secured resource
+  * a 'Real Case' scenario, where we can log in, and obtain or use secured endpoints (provided by the Auth server and by a Custom server we set up)
+
+The Step-By-Step guide supports using different providers (Authorization Servers) by just adding (or uncommenting) the corresponding entries in the static/*spa*/js/configs.js.
+
+### The 'Step-by-Step' OAuth Client with PKCE page
+After running the Spring Boot Application (a simple *mvn spring-boot:run* command will be enough), we can browse to *http://localhost:8080/pkce-stepbystep/index.html* and follow the steps to find out what it takes to obtain an access token using the Authorization Code with PKCE Flow.
+
+When prompted the login form, we might need to create a user for our Application first.
+
+### The 'Real-Case' OAuth Client with PKCE page
+To use all the features contained in the *http://localhost:8080/pkce-realcase/index.html* page, we'll need to first start the resource server (SPA/oauth-resource-server-auth0).
+
+In this page, we can:
+  * List the resources in our resource server (public, no permissions needed)
+  * Add resources (we're requested the permissions to do that when logging in. For simplicity sake, we just request the existing 'profile' scope)
+  * Remove resources (we actually can't accomplish this task, because the resource server requires the application to have permissions that were not included in the existing scopes)
 
 ## Relevant Articles: 
 - [Spring REST API + OAuth2 + AngularJS](http://www.baeldung.com/rest-api-spring-oauth2-angularjs)
