+++ SPA/clients-js-only-react/pom.xml
