+++ SPA/clients-js-only-react/src/main/java/com/baeldung/clientjsonlyreact/ClientJsOnlyReactApplication.java
