+++ SPA/clients-js-only-react/src/main/resources/static/common/images/default-profile.png
