+++ SPA/clients-js-only-react/src/main/resources/static/common/js/code_challenge_functions.js
