+++ SPA/clients-js-only-react/src/main/resources/static/common/js/components/Spinner.js
