+++ SPA/clients-js-only-react/src/main/resources/static/pkce-realcase/js/components/pages/ListingItem.js
@@ -1,0 +1,10 @@
+// Listing Item
+const ListingItem = ({ id, value, onDeleteFn }) => (
+  <div key={id} className="listing-item">
+  <div style={{backgroundColor:value}} className="color-demo"></div>
+    <div className="color-name">{value}</div>
+    <div onClick={onDeleteFn.bind(this,id)} className="fas fa-trash-alt"></div>
+  </div>
+)
+
+window.ListingItem = ListingItem;
