+++ SPA/clients-js-only-react/src/main/resources/static/pkce-stepbystep/js/components/steps/Step2.js
