+++ SPA/clients-js-only-react/src/test/java/com/baeldung/clientjsonlyreact/ClientJsOnlyReactApplicationTests.java
