+++ SPA/oauth-resource-server-auth0/src/main/java/com/baeldung/resourceserverauth0/OauthResourceServerAuth0Application.java
@@ -1,0 +1,14 @@
+package com.baeldung.resourceserverauth0;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OauthResourceServerAuth0Application {
+
+	public static void main(String[] args) {
+		SpringApplication.run(OauthResourceServerAuth0Application.class, args);
+	}
+
+}
+
