+++ SPA/oauth-resource-server-auth0/src/main/java/com/baeldung/resourceserverauth0/dao/ColorRepository.java
@@ -1,0 +1,9 @@
+package com.baeldung.resourceserverauth0.dao;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.baeldung.resourceserverauth0.model.Color;
+
+public interface ColorRepository extends CrudRepository<Color, Long> {
+
+}
