+++ SPA/oauth-resource-server-auth0/src/main/resources/data.sql
@@ -1,0 +1,8 @@
+insert into color
+values(1001, 'red');
+
+insert into color
+values(1002, 'green');
+
+insert into color
+values(1003, 'blue');
