+++ SPA/oauth-resource-server-auth0/src/test/java/com/baeldung/resourceserverauth0/OauthResourceServerAuth0ApplicationTests.java
@@ -1,0 +1,17 @@
+package com.baeldung.resourceserverauth0;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class OauthResourceServerAuth0ApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
+
